fix: refactor demo-app
moved enums and structures to different files
fixed indents in classes/models
