+++ REES46/Classes/Model/CartItem.swift
@@ -1,16 +1,16 @@
 import Foundation
 
 public struct CartItem {
-    var productId: String
-    var quantity: Int = 1
-    
-    public init(productId: String, quantity: Int = 1) {
-        self.productId = productId
-        self.quantity = quantity
-    }
-    
-    public init(json: [String: Any]) {
-        self.productId = json["uniqid"] as? String ?? ""
-        self.quantity = json["quantity"] as? Int ?? 1
-    }
+  var productId: String
+  var quantity: Int = 1
+  
+  public init(productId: String, quantity: Int = 1) {
+    self.productId = productId
+    self.quantity = quantity
+  }
+  
+  public init(json: [String: Any]) {
+    self.productId = json["uniqid"] as? String ?? ""
+    self.quantity = json["quantity"] as? Int ?? 1
+  }
 }
