+++ REES46/Classes/Model/Reccomend/RecommendConstants.swift
@@ -1,0 +1,46 @@
+import Foundation
+
+struct Constants {
+  struct JSONKeys {
+    static let recommends = "recommends"
+    static let title = "title"
+    static let locations = "locations"
+    static let id = "id"
+    static let barcode = "barcode"
+    static let name = "name"
+    static let brand = "brand"
+    static let model = "model"
+    static let description = "description"
+    static let imageUrl = "image_url"
+    static let picture = "picture"
+    static let url = "url"
+    static let deeplinkIos = "deeplink_ios"
+    static let price = "price"
+    static let priceFormatted = "price_formatted"
+    static let priceFull = "price_full"
+    static let priceFullFormatted = "price_full_formatted"
+    static let oldPrice = "oldprice"
+    static let oldPriceFormatted = "oldprice_formatted"
+    static let oldPriceFull = "oldprice_full"
+    static let oldPriceFullFormatted = "oldprice_full_formatted"
+    static let currency = "currency"
+    static let salesRate = "sales_rate"
+    static let relativeSalesRate = "relative_sales_rate"
+    static let discount = "discount"
+    static let rating = "rating"
+    static let image_url_resized = "image_url_resized"
+    static let categories = "categories"
+    static let withLocations = "with_locations"
+    static let extended = "extended"
+    static let fashionOriginalSizes = "fashion_original_sizes"
+    static let fashionSizes = "fashion_sizes"
+    static let fashionColors = "fashion_colors"
+    static let params = "params"
+  }
+  
+  struct RecommendedBy {
+    static let recommendedBy = "recommended_by"
+    static let recommendedCode = "recommended_code"
+    static let webPushDigestCode = "web_push_digest_code"
+  }
+}
