+++ REES46/Classes/Model/SdkConstants.struct.swift
@@ -1,36 +1,36 @@
 import Foundation
 
 struct SdkConstants {
-    static let brands: String = "brands"
-    static let categories: String = "categories"
-    static let categoryLimit: String = "category_limit"
-    static let colors: String = "colors"
-    static let defaultTimeout: Double = 1.0
-    static let deviceIdKey = "device_id"
-    static let deviceToken = "device_token"
-    static let disableClarification: String = "no_clarification"
-    static let email: String = "email"
-    static let exclude: String = "exclude"
-    static let extended: String = "extended"
-    static let fashionSizes: String = "fashion_sizes"
-    static let filters: String = "filters"
-    static let fullSearch: String = "full_search"
-    static let limit: String = "limit"
-    static let locations: String = "locations"
-    static let noClarificationValue: String = "1"
-    static let offset: String = "offset"
-    static let priceMax: String = "price_max"
-    static let priceMin: String = "price_min"
-    static let searchPath: String = "search"
-    static let searchQuery: String = "search_query"
-    static let seance = "seance"
-    static let segment: String = "segment"
-    static let shopId = "shop_id"
-    static let tz = "tz"
-    static let sid = "sid"
-    static let did = "did"
-    static let sortBy: String = "sort_by"
-    static let sortDir: String = "sort_dir"
-    static let type: String = "type"
-    static let keychainDid: String = "keychainDid"
+  static let brands: String = "brands"
+  static let categories: String = "categories"
+  static let categoryLimit: String = "category_limit"
+  static let colors: String = "colors"
+  static let defaultTimeout: Double = 1.0
+  static let deviceIdKey = "device_id"
+  static let deviceToken = "device_token"
+  static let disableClarification: String = "no_clarification"
+  static let email: String = "email"
+  static let exclude: String = "exclude"
+  static let extended: String = "extended"
+  static let fashionSizes: String = "fashion_sizes"
+  static let filters: String = "filters"
+  static let fullSearch: String = "full_search"
+  static let limit: String = "limit"
+  static let locations: String = "locations"
+  static let noClarificationValue: String = "1"
+  static let offset: String = "offset"
+  static let priceMax: String = "price_max"
+  static let priceMin: String = "price_min"
+  static let searchPath: String = "search"
+  static let searchQuery: String = "search_query"
+  static let seance = "seance"
+  static let segment: String = "segment"
+  static let shopId = "shop_id"
+  static let tz = "tz"
+  static let sid = "sid"
+  static let did = "did"
+  static let sortBy: String = "sort_by"
+  static let sortDir: String = "sort_dir"
+  static let type: String = "type"
+  static let keychainDid: String = "keychainDid"
 }
