+++ REES46/Classes/Profile/Service/Model/ProfileGender.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum Gender: String{
+  case male
+  case female
+}
+
